It is now possible to sort thumbnails by attributes in Gkofoto. This version of Gkofoto request thumbnails of the size that are going to be shown to the user, from the image cache, instead as such before, when it tried to reuse existing images in the cache. Hopefully this can make GKofoto a little faster.

+++ src/gnome/imageview.py
@@ -64,8 +64,8 @@ class ImageView(gtk.ScrolledWindow):
             pixBufResized = self._pixBuf
         else:
             zoomMultiplicator = pow(self._ZOOMFACTOR, self._wantedZoom)
-            wantedWidth = self._pixBuf.get_width() * zoomMultiplicator 
-            wantedHeight = self._pixBuf.get_height() * zoomMultiplicator 
+            wantedWidth = int(self._pixBuf.get_width() * zoomMultiplicator)
+            wantedHeight = int(self._pixBuf.get_height() * zoomMultiplicator)
             if min(wantedWidth, wantedHeight) < self._MIN_IMAGE_SIZE:
                 # Too small image size
                 return
